Implement StorageProviderHelper in the runtime
